Improve search result sorting and add subtitles to results (#212)
## 🐛 Issue

<!-- Link to the issue in Github or Basecamp this PR is addressing. If there is no related issue or related pull request, consider briefly describing the problem or enhancement being addressed. -->

Since we have two indices, search results always return with web page results first, and content item results second.

This makes it hard to the user to understand which results to look at, and without any other result context (like summaries) the quality of the search results feel very poor.

## ✏️ Solution

<!--
Describe your changes, and why you're making them. If there's something novel or complex about your approach, you can call it out here. 
-->

1. Combine multi-index results into a single array. This is hacky, and isn't a permanent solution. Best case - the results end up in a much better order. Worse case - they end up in a similar order to they are already. 

2. Add summary/subtitles to results

3. Remove "Search history" section if you have search results

As a follow up, @vinnyjth is going to document work to be done on the shovel to create a new, combined index that we can start tweaking to more dramatically improve search results.

Also, we will need to pursue a minor (hopefully) refactor for search on the web. Currently, all search is done through the autocomplete API on Algolia, not the instantsearch api. This means that features like snippeting is not available, and we also can not use any of Algolia's out of the box react components. Not using algolia best practices, unfortunately.

## 🔬 To Test

<!--
With only the context in this description, how would a developer from outside Apollos setup and validate your change? 
-->

1. Boot up web embeds
2. Test search

## 📸 Screenshots

| | Before | After |
| --- | --- | --- |
| **Best Case:** "Take Heart" content item is now listed first| ![Screenshot 2024-04-09 at 9.01.24 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/Iqv8SYVwCDX0dUbjBiDz/c8b282f9-faef-4d6e-9f31-b8de18338b4e.png) |  ![Screenshot 2024-04-09 at 9.00.07 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/Iqv8SYVwCDX0dUbjBiDz/a1b8a106-7f5c-4956-abab-847157fe0aaa.png) | 
| **Worse Case:** The "Campus Locations" result is the best result, but still shown last| ![Screenshot 2024-04-09 at 9.02.50 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/Iqv8SYVwCDX0dUbjBiDz/c9c142b1-11a9-4fb1-9cb6-58c556508526.png) | ![Screenshot 2024-04-09 at 9.00.36 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/Iqv8SYVwCDX0dUbjBiDz/7b32df41-eb3a-43e2-ab96-799c6d36955d.png) |